renaming notebook to nb
