+++ tools/test_website.ts
@@ -1,4 +1,4 @@
 // Tests that only run in the browser.
+import "../website/nb_test";
 import "../website/nb_transpiler_test";
-import "../website/notebook_test";
 import "./test_isomorphic";
